restructure folders to avoid possible namespace collision
source files are restructured according to coding guidelines
