+++ examples/ultrahdr_app.cpp
@@ -27,9 +27,9 @@
 #include <fstream>
 #include <iostream>
 
-#include "ultrahdrcommon.h"
-#include "gainmapmath.h"
-#include "jpegr.h"
+#include "ultrahdr/ultrahdrcommon.h"
+#include "ultrahdr/gainmapmath.h"
+#include "ultrahdr/jpegr.h"
 
 using namespace ultrahdr;
