+++ lib/include/ultrahdr/gainmapmath.h
