+++ lib/include/ultrahdr/jpegdecoderhelper.h
