+++ lib/include/ultrahdr/jpegencoderhelper.h
