+++ lib/include/ultrahdr/jpegr.h
@@ -19,9 +19,9 @@
 
 #include <cfloat>
 
-#include "ultrahdr.h"
-#include "jpegdecoderhelper.h"
-#include "jpegencoderhelper.h"
+#include "ultrahdr/ultrahdr.h"
+#include "ultrahdr/jpegdecoderhelper.h"
+#include "ultrahdr/jpegencoderhelper.h"
 
 namespace ultrahdr {
