+++ lib/include/ultrahdr/multipictureformat.h
@@ -33,10 +33,10 @@
 #define Endian_SwapBE16(n) (n)
 #endif
 
-#include "ultrahdr.h"
-#include "jpegr.h"
-#include "gainmapmath.h"
-#include "jpegrutils.h"
+#include "ultrahdr/ultrahdr.h"
+#include "ultrahdr/jpegr.h"
+#include "ultrahdr/gainmapmath.h"
+#include "ultrahdr/jpegrutils.h"
 
 namespace ultrahdr {
