+++ lib/src/jpegdecoderhelper.cpp
@@ -19,9 +19,9 @@
 
 #include <cstring>
 
-#include "ultrahdrcommon.h"
-#include "ultrahdr.h"
-#include "jpegdecoderhelper.h"
+#include "ultrahdr/ultrahdrcommon.h"
+#include "ultrahdr/ultrahdr.h"
+#include "ultrahdr/jpegdecoderhelper.h"
 
 using namespace std;
