+++ lib/src/jpegencoderhelper.cpp
@@ -18,9 +18,9 @@
 #include <memory>
 #include <string>
 
-#include "ultrahdrcommon.h"
-#include "ultrahdr.h"
-#include "jpegencoderhelper.h"
+#include "ultrahdr/ultrahdrcommon.h"
+#include "ultrahdr/ultrahdr.h"
+#include "ultrahdr/jpegencoderhelper.h"
 
 namespace ultrahdr {
