+++ tests/jpegr_test.cpp
@@ -24,9 +24,9 @@
 #include <fstream>
 #include <iostream>
 
-#include "ultrahdrcommon.h"
-#include "jpegr.h"
-#include "jpegrutils.h"
+#include "ultrahdr/ultrahdrcommon.h"
+#include "ultrahdr/jpegr.h"
+#include "ultrahdr/jpegrutils.h"
 
 //#define DUMP_OUTPUT
